xml/parsing/manakai/xmldecls-1.dat

#data
<?xml?><a/>
#errors
1;3;attribute missing:version;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml ?><a/>
#errors
1;3;attribute missing:version;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml version="1.0"?><a/>
#errors
#xml-version
1.0
#document
| <{} a>

#data
<?xml version   =  "1.0"?><a/>
#errors
#xml-version
1.0
#document
| <{} a>

#data
<?xml Version="1.0"?><a/>
#errors
1;7;attribute missing:version;;;m
1;7;bogus XML declaration;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml version=1.0?><a/>
#errors
1;7;attribute missing:version;;;m
1;7;bogus XML declaration;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml version="1.0"  ?><a/>
#errors
#xml-version
1.0
#document
| <{} a>

#data
<?xml version='1.0'?> <a/>
#errors
#xml-version
1.0
#document
| <{} a>

#data
<?xml version="1.0'?><a/>
#errors
1;7;attribute missing:version;;;m
1;7;bogus XML declaration;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml version="1.1"?><a/>
#errors
1;16;xml:version:not 1.0;;1.1;s
#xml-version
1.1
#document
| <{} a>

#data
<?xml version="1.2"?><a/>
#errors
1;16;xml:version:not 1.0;;1.2;s
#xml-version
1.2
#document
| <{} a>

#data
<?xml version="'1.0'"?><a/>
#errors
1;16;XML version:syntax error;;'1.0';m
#xml-version
'1.0'
#document
| <{} a>

#data
<?xml version="unknown"?><a/>
#errors
1;16;XML version:syntax error;;unknown;m
#xml-version
unknown
#document
| <{} a>

#data
<?xml version="1.1" unknown?><a/>
#errors
1;16;xml:version:not 1.0;;1.1;s
1;21;bogus XML declaration;;;m
#xml-version
1.1
#document
| <{} a>

#data
<?xml version?><a/>
#errors
1;7;attribute missing:version;;;m
1;7;bogus XML declaration;;;m
#xml-version
1.0
#document
| <{} a>

#data
<?xml version=""?><a/>
#errors
1;16;XML version:syntax error;;;m
#xml-version
#document
| <{} a>

#data
<?xml encoding=""?><a/>
#errors
1;7;attribute missing:version;;;m
1;17;XML encoding:syntax error;;;m
#xml-version
1.0
#xml-encoding
#document
| <{} a>

#data
<?xml version="1.0"encoding="a"?><a/>
#errors
1;20;no space before attr name;;;m
#xml-version
1.0
#xml-encoding
a
#document
| <{} a>

#data
<?xml version="1.0" encoding=""?><a/>
#errors
1;31;XML encoding:syntax error;;;m
#xml-version
1.0
#xml-encoding
#document
| <{} a>

#data
<?xml version="1.0" encoding="us-ascii"?><a/>
#errors
#xml-version
1.0
#xml-encoding
us-ascii
#document
| <{} a>

#data
<?xml version="1.0" encoding="UtF-8"?><a/>
#errors
#xml-version
1.0
#xml-encoding
UtF-8
#document
| <{} a>

#data
<?xml version="1.0" encoding='shift_Jis'?><a/>
#errors
#xml-version
1.0
#xml-encoding
shift_Jis
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-16'?><a/>
#errors
1;21;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#document
| <{} a>

#data
<?xml version="1.0" encoding  =   "UTF-32BE"?><a/>
#errors
#xml-version
1.0
#xml-encoding
UTF-32BE
#document
| <{} a>

#data
<?xml version="1.0" encoding="hz"     ?><a/>
#errors
#xml-version
1.0
#xml-encoding
hz
#document
| <{} a>

#data
<?xml version="1.0" encoding=iso-8859-1?><a/>
#errors
1;21;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#document
| <{} a>

#data
<?xml version="1.0" encoding="windows-1252" a?><a/>
#errors
1;45;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding
windows-1252
#document
| <{} a>

#data escaped
<?xml version="1.0" encoding="\u4e00"?><a/>
#errors escaped
1;31;XML encoding:syntax error;;\u4E00;m
#xml-version
1.0
#xml-encoding escaped
\u4e00
#document
| <{} a>

#data
<?xml version="1.0" encoding=" windows-31j "?><a/>
#errors
1;31;XML encoding:syntax error;; windows-31j ;m
#xml-version
1.0
#xml-encoding escaped
 windows-31j\u0020
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-8" standalone="yes"?><a/>
#errors
#xml-version
1.0
#xml-encoding
utf-8
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-8" standalone='yes'?><a/>
#errors
#xml-version
1.0
#xml-encoding
utf-8
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-8" standalone  =   "yes"?><a/>
#errors
#xml-version
1.0
#xml-encoding
utf-8
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-8" standalone  =  "no"?><a/>
#errors
#xml-version
1.0
#xml-encoding
utf-8
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0" encoding="utf-8" standalone="yes" ?><a/>
#errors
#xml-version
1.0
#xml-encoding
utf-8
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0"  standalone="yes"?><a/>
#errors
#xml-version
1.0
#xml-encoding null
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0"  standalone="no"?><a/>
#errors
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml  standalone="yes"?><a/>
#errors
1;8;attribute missing:version;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone true
#document
| <{} a>

#data
<?xml  standalone="no"?><a/>
#errors
1;8;attribute missing:version;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0"standalone="yes"?><a/>
#errors
1;20;no space before attr name;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" standalone="yes'?><a/>
#errors
1;21;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0" standalone="yes" encoding="utf-8"?><a/>
#errors
1;38;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" standalone="yes" unknown=""?><a/>
#errors
1;38;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone true
#document
| <{} a>

#data
<?xml version="1.0" standalone="true"?><a/>
#errors
1;33;XML standalone:syntax error;;true;m
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0" standAlone="yes"?><a/>
#errors
1;21;bogus XML declaration;;;m
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0" standalone="YES"?><a/>
#errors
1;33;XML standalone:syntax error;;YES;m
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>

#data
<?xml version="1.0"?>
<?xml version="1.0" standalone="yes"?><a/>
#errors
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <?xml version="1.0" standalone="yes"?>
| <{} a>

#data
<?xml version="1.1" encoding="utf-16"?>
<?xml version="1.0" encoding="utf-8"?><a/>
#errors
1;16;xml:version:not 1.0;;1.1;s
#xml-version
1.1
#xml-encoding
utf-16
#xml-standalone false
#document
| <?xml version="1.0" encoding="utf-8"?>
| <{} a>

#data
<!DOCTYPE HTML><?xml version="1.0" standalone="yes"?><a/>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <!DOCTYPE HTML>
| <?xml version="1.0" standalone="yes"?>
| <{} a>

#data
 <?xml version="1.0" standalone="yes"?><a/>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <?xml version="1.0" standalone="yes"?>
| <{} a>

#data escaped
\u000a<?xml version="1.0" standalone="yes"?><a/>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <?xml version="1.0" standalone="yes"?>
| <{} a>

#data
<?XML version="1.0" standalone="yes"?><a/>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <?XML version="1.0" standalone="yes"?>
| <{} a>

#data
<?xml1 version="1.0" standalone="yes"?><a/>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <?xml1 version="1.0" standalone="yes"?>
| <{} a>

#data
<a><?xml version="1.0" standalone="yes"?></a>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>
|   <?xml version="1.0" standalone="yes"?>

#data
<a/><?xml version="1.0" standalone="yes"?>
#errors
1;1;no XML decl;;;s
#xml-version
1.0
#xml-encoding null
#xml-standalone false
#document
| <{} a>
| <?xml version="1.0" standalone="yes"?>
